Save empty gt also as uint8

+++ src/intrab/experiments.py
@@ -91,7 +91,7 @@ def run_experiments(
             if np.all(binary_gt == 0):
                 logger.debug(f"Skipping {gt_path} missing segmentation for {target}")
                 img = nib.load(gt_path)
-                empty_gt = nib.Nifti1Image(binary_gt.astype(np.float32), img.affine)
+                empty_gt = nib.Nifti1Image(binary_gt.astype(np.uint8), img.affine)
                 empty_gt.to_filename(results_dir / prompter.name / target_name / base_name)
                 continue
